+++ packages/dash-core-components/src/index.js
@@ -22,6 +22,7 @@ import Tabs from './components/Tabs.react';
 import Tab from './components/Tab.react';
 import Store from './components/Store.react';
 import LogoutButton from './components/LogoutButton.react';
+import Clipboard from './components/Clipboard.react';
 
 import 'react-dates/lib/css/_datepicker.css';
 import './components/css/react-dates@20.1.0-fix.css';
@@ -49,5 +50,6 @@ export {
     Upload,
     Store,
     LogoutButton,
-    Download
+    Download,
+    Clipboard
 };